Move check_stack, stack_checker, get_debug, set_debug to ll_api
