+++ pynest/nest/hl_api.py
@@ -34,38 +34,37 @@ for name in os.listdir(os.path.join(os.path.dirname(__file__), "lib")):
         exec("from .lib.{0} import *".format(name[:-3]))
 
 # With '__all__' we provide an explicit index of the package. Without any
-# imported submodules and any redundant functions we could achieve shorter
-# time of importing 'nest'.
+# imported submodules and any redundant functions we could minimize list.
 __all__ = [
-    'BeginSubnet',  # deprecated?
+    'BeginSubnet',  # deprecated
     'CGConnect',
     'CGParse',
     'CGSelectImplementation',
-    'ChangeSubnet',  # deprecated?
+    'ChangeSubnet',  # deprecated
     'Cleanup',
     'Connect',
     'ConnectionRules',
     'CopyModel',
     'Create',
-    'CurrentSubnet',  # deprecated?
-    'DataConnect',  # deprecated?
+    'CurrentSubnet',  # deprecated
+    'DataConnect',  # deprecated
     'DisableStructuralPlasticity',
     'Disconnect',
     'DisconnectOneToOne',
     'EnableStructuralPlasticity',
-    'EndSubnet',  # deprecated?
-    'GetChildren',  # deprecated?
+    'EndSubnet',  # deprecated
+    'GetChildren',  # deprecated
     'GetConnections',
     'GetDefaults',
     'GetKernelStatus',
-    'GetLeaves',  # deprecated?
-    'GetLID',  # deprecated?
-    'GetNetwork',  # deprecated?
-    'GetNodes',  # deprecated?
+    'GetLeaves',  # deprecated
+    'GetLID',  # deprecated
+    'GetNetwork',  # deprecated
+    'GetNodes',  # deprecated
     'GetStatus',
     'GetStructuralPlasticityStatus',
     'Install',
-    'LayoutNetwork',  # deprecated?
+    'LayoutNetwork',  # deprecated
     'Models',
     'NumProcesses',
     'Prepare',
@@ -82,37 +81,9 @@ __all__ = [
     'SetStatus',
     'SetStructuralPlasticityStatus',
     'Simulate',
-    # 'SuppressedDeprecationWarning',  # redundant?
-    # 'Template',
     'authors',
-    # 'broadcast',  # redundant?
-    # 'catching_sli_run',  # redundant?
-    # 'check_stack',  # redundant?
-    # 'deprecated',  # redundant?
-    # 'functools',  # redundant?
-    # 'get_argv',  # redundant?
-    # 'get_debug',  # redundant?
-    # 'get_help_filepath',  # redundant?
-    # 'get_unistring_type',  # redundant?
-    # 'get_verbosity',
-    # 'get_wrapped_text',  # redundant?
     'help',
     'helpdesk',
-    # 'is_coercible_to_sli_array',  # redundant?
-    # 'is_iterable',  # redundant?
-    # 'is_literal',  # redundant?
-    # 'is_sequence_of_connections',  # redundant?
-    # 'is_sequence_of_gids',  # redundant?
-    # 'is_string',  # redundant?
-    # 'load_help',  # redundant?
-    # 'message',  # redundant?
-    # 'model_deprecation_warning',  # redundant?
-    # 'name',  # redundant?
-    # 'set_debug',  # redundant?
-    # 'set_verbosity',
-    # 'show_deprecation_warning',  # redundant?
-    # 'show_help_with_pager',  # redundant?
-    # 'stack_checker',  # redundant?
     'sysinfo',
     'version',
 ]
