+++ pynest/nest/lib/hl_api_helper.py
@@ -25,7 +25,6 @@ API of the PyNEST wrapper.
 """
 
 import warnings
-import inspect
 import json
 import functools
 import textwrap
@@ -40,29 +39,25 @@ from ..ll_api import *
 from .. import pynestkernel as kernel
 
 __all__ = [
-    'get_wrapped_text',
-    'show_deprecation_warning',
+    'broadcast',
     'deprecated',
+    'get_help_filepath',
     'get_unistring_type',
-    'uni_str',
-    'is_literal',
-    'is_string',
-    'get_debug',
-    'set_debug',
-    'stack_checker',
-    'check_stack',
-    'is_iterable',
+    'get_verbosity',
+    'get_wrapped_text',
     'is_coercible_to_sli_array',
+    'is_iterable',
+    'is_literal',
     'is_sequence_of_connections',
     'is_sequence_of_gids',
-    'broadcast',
-    'get_help_filepath',
+    'is_string',
     'load_help',
-    'show_help_with_pager',
-    'get_verbosity',
-    'set_verbosity',
     'model_deprecation_warning',
+    'set_verbosity',
+    'show_deprecation_warning',
+    'show_help_with_pager',
     'SuppressedDeprecationWarning',
+    'uni_str',
 ]
 
 
@@ -197,108 +192,6 @@ def is_string(obj):
         True if obj is a unicode string
     """
     return isinstance(obj, uni_str)
-
-__debug = False
-
-
-def get_debug():
-    """Return the current value of the debug flag of the high-level API.
-
-    Returns
-    -------
-    bool:
-        current value of the debug flag
-    """
-
-    global __debug
-    return __debug
-
-
-def set_debug(dbg=True):
-    """Set the debug flag of the high-level API.
-
-    Parameters
-    ----------
-    dbg : bool, optional
-        Value to set the debug flag to
-    """
-
-    global __debug
-    __debug = dbg
-
-
-def stack_checker(f):
-    """Decorator to add stack checks to functions using PyNEST's
-    low-level API.
-
-    This decorator works only on functions. See
-    check_stack() for the generic version for functions and
-    classes.
-
-    Parameters
-    ----------
-    f : function
-        Function to decorate
-
-    Returns
-    -------
-    function:
-        Decorated function
-
-    Raises
-    ------
-    kernel.NESTError
-    """
-
-    @functools.wraps(f)
-    def stack_checker_func(*args, **kwargs):
-        if not get_debug():
-            return f(*args, **kwargs)
-        else:
-            sr('count')
-            stackload_before = spp()
-            result = f(*args, **kwargs)
-            sr('count')
-            num_leftover_elements = spp() - stackload_before
-            if num_leftover_elements != 0:
-                eargs = (f.__name__, num_leftover_elements)
-                etext = "Function '%s' left %i elements on the stack."
-                raise kernel.NESTError(etext % eargs)
-            return result
-
-    return stack_checker_func
-
-
-def check_stack(thing):
-    """Convenience wrapper for applying the stack_checker decorator to
-    all class methods of the given class, or to a given function.
-
-    If the object cannot be decorated, it is returned unchanged.
-
-    Parameters
-    ----------
-    thing : function or class
-        Description
-
-    Returns
-    -------
-    function or class
-        Decorated function or class
-
-    Raises
-    ------
-    ValueError
-    """
-
-    if inspect.isfunction(thing):
-        return stack_checker(thing)
-    elif inspect.isclass(thing):
-        for name, mtd in inspect.getmembers(thing, predicate=inspect.ismethod):
-            if name.startswith("test_"):
-                setattr(thing, name, stack_checker(mtd))
-        return thing
-    else:
-        raise ValueError("unable to decorate {0}".format(thing))
 
 
 def is_iterable(seq):
