+++ pynest/nest/lib/hl_api_subnets.py
@@ -30,16 +30,16 @@ from .hl_api_nodes import Create
 from .hl_api_info import GetStatus, SetStatus
 
 __all__ = [
-    'PrintNetwork',
-    'CurrentSubnet',
-    'ChangeSubnet',
-    'GetLeaves',
-    'GetNodes',
-    'GetChildren',
-    'GetNetwork',
     'BeginSubnet',
+    'ChangeSubnet',
+    'CurrentSubnet',
     'EndSubnet',
+    'GetChildren',
+    'GetLeaves',
+    'GetNetwork',
+    'GetNodes',
     'LayoutNetwork',
+    'PrintNetwork',
 ]
