+++ pynest/nest/tests/test_all.py
@@ -98,10 +98,10 @@ def suite():
 
 if __name__ == "__main__":
 
-    debug = nest.hl_api.get_debug()
-    nest.hl_api.set_debug(True)
+    debug = nest.ll_api.get_debug()
+    nest.ll_api.set_debug(True)
 
     runner = unittest.TextTestRunner(verbosity=2)
     runner.run(suite())
 
-    nest.hl_api.set_debug(debug)
+    nest.ll_api.set_debug(debug)
