+++ pynest/nest/tests/test_connect_all_to_all.py
@@ -27,7 +27,7 @@ from . import test_connect_helpers as hf
 from .test_connect_parameters import TestParams
 
 
-@hf.nest.hl_api.check_stack
+@hf.nest.ll_api.check_stack
 class TestAllToAll(TestParams):
 
     # specify connection pattern
