+++ pynest/nest/tests/test_connect_array_fixed_indegree.py
@@ -29,7 +29,7 @@ import nest
 import numpy
 
 
-@nest.hl_api.check_stack
+@nest.ll_api.check_stack
 class ConnectArrayFixedIndegreeTestCase(unittest.TestCase):
     """Tests of connections with fixed indegree and parameter arrays"""
