+++ pynest/nest/tests/test_current_recording_generators.py
@@ -28,7 +28,7 @@ import unittest
 import nest
 
 
-@nest.hl_api.check_stack
+@nest.ll_api.check_stack
 class CurrentRecordingGeneratorTestCase(unittest.TestCase):
     """
     Test if currents from generators are recorded properly. Specifically:
