+++ pynest/nest/tests/test_dataconnect.py
@@ -27,7 +27,7 @@ import unittest
 import nest
 
 
-@nest.hl_api.check_stack
+@nest.ll_api.check_stack
 class DataConnectTestCase(unittest.TestCase):
     """Find connections and test if values can be set."""
