+++ pynest/nest/tests/test_errors.py
@@ -27,7 +27,7 @@ import unittest
 import nest
 
 
-@nest.hl_api.check_stack
+@nest.ll_api.check_stack
 class ErrorTestCase(unittest.TestCase):
     """Tests if errors are handled correctly"""
