+++ pynest/nest/tests/test_parrot_neuron_ps.py
@@ -36,7 +36,7 @@ def _round_up(simtime):
     return res * math.ceil(float(simtime) / float(res))
 
 
-@nest.hl_api.check_stack
+@nest.ll_api.check_stack
 class ParrotNeuronPSTestCase(unittest.TestCase):
     """Check parrot_neuron spike repetition properties"""
 
@@ -108,7 +108,7 @@ class ParrotNeuronPSTestCase(unittest.TestCase):
             "Parrot neuron failed to correctly repeat multiple spikes."
 
 
-@nest.hl_api.check_stack
+@nest.ll_api.check_stack
 class ParrotNeuronPSPoissonTestCase(unittest.TestCase):
     """Check parrot_neuron spike repetition properties"""
 
@@ -166,7 +166,7 @@ class ParrotNeuronPSPoissonTestCase(unittest.TestCase):
             "parrot_neuron adds spikes."
 
 
-@nest.hl_api.check_stack
+@nest.ll_api.check_stack
 class ParrotNeuronPSSTDPTestCase(unittest.TestCase):
     """
     Check STDP protocol between two parrot_neurons_ps connected by a
