+++ pynest/nest/tests/test_quantal_stp_synapse.py
@@ -26,7 +26,7 @@ import numpy
 import unittest
 
 
-@nest.hl_api.check_stack
+@nest.ll_api.check_stack
 class QuantalSTPSynapseTestCase(unittest.TestCase):
     """Compare quantal_stp_synapse with its deterministic equivalent."""
