+++ pynest/nest/tests/test_rate_neuron.py
@@ -30,7 +30,7 @@ import unittest
 import numpy as np
 
 
-@nest.hl_api.check_stack
+@nest.ll_api.check_stack
 class RateNeuronTestCase(unittest.TestCase):
 
     """Check rate_neuron"""
