+++ pynest/nest/tests/test_rate_neuron_communication.py
@@ -32,7 +32,7 @@ def H(x):
     return 0.5 * (np.sign(x) + 1.)
 
 
-@nest.hl_api.check_stack
+@nest.ll_api.check_stack
 class RateNeuronCommunicationTestCase(unittest.TestCase):
 
     """Check rate_neuron"""
