+++ pynest/nest/tests/test_siegert_neuron.py
@@ -28,7 +28,7 @@ import numpy as np
 HAVE_GSL = nest.ll_api.sli_func("statusdict/have_gsl ::")
 
 
-@nest.hl_api.check_stack
+@nest.ll_api.check_stack
 @unittest.skipIf(not HAVE_GSL, 'GSL is not available')
 class SiegertNeuronTestCase(unittest.TestCase):
     """
