+++ pynest/nest/tests/test_vogels_sprekeler_synapse.py
@@ -26,7 +26,7 @@ import unittest
 from math import exp
 
 
-@nest.hl_api.check_stack
+@nest.ll_api.check_stack
 class VogelsSprekelerConnectionTestCase(unittest.TestCase):
 
     """Check vogels_sprekeler_synapse model properties."""
