+++ topology/pynest/hl_api.py
@@ -66,7 +66,7 @@ using the same dictionary to specify both connections.
 import nest
 from .ll_api import topology_func
 
-# With '__all__' we provide an explicit index of this submodule. 
+# With '__all__' we provide an explicit index of this submodule.
 __all__ = [
     'ConnectLayers',
     'CreateLayer',
